fix: 페이지 컴포넌트에서 수행하던 ai 채점 작업을 하위 컴포넌트로 이관
